# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CEREBRO-RED v2 API",
    "description": "\n    # Autonomous Local LLM Red Teaming Suite - Research Edition\n    \n    ## Overview\n    CEREBRO-RED v2 implements the PAIR (Prompt Automatic Iterative Refinement) algorithm\n    for automated vulnerability discovery in Large Language Models.\n    \n    ## Features\n    - **Multi-Provider LLM Support**: Ollama, Azure OpenAI, OpenAI\n    - **8 Attack Strategies**: Obfuscation, Context Flooding, Roleplay Injection, etc.\n    - **LLM-as-a-Judge**: Semantic evaluation with 7 criteria\n    - **Real-Time Progress**: WebSocket streaming for live updates\n    - **Research-Grade Telemetry**: JSONL audit logs for analysis\n    \n    ## Authentication\n    All endpoints (except `/health`, `/docs`, `/metrics`) require API key authentication via `X-API-Key` header.\n    \n    ## Rate Limiting\n    Default: 60 requests per minute per IP address.\n    \n    ## References\n    - PAIR Paper: https://arxiv.org/abs/2310.08419\n    - GitHub: https://github.com/your-org/cerebro-red-v2\n    ",
    "version": "2.0.0",
    "contact": {
      "name": "CEREBRO-RED v2 Support",
      "url": "https://github.com/your-org/cerebro-red-v2/issues",
      "email": "support@cerebro-red.example.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "paths": {
...
    "/api/v1/experiments": {
      "post": {
        "tags": [
          "experiments"
        ],
        "summary": "Create Experiment",
        "description": "Create a new experiment with enhanced error logging.",
        "operationId": "create_experiment_api_v1_experiments_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExperimentConfig"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "experiments"
        ],
        "summary": "List Experiments",
        "description": "List all experiments with pagination.",
        "operationId": "list_experiments_api_v1_experiments_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Page Size"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/experiments/{experiment_id}": {
      "get": {
        "tags": [
          "experiments"
        ],
        "summary": "Get Experiment",
        "description": "Get experiment details by ID.",
        "operationId": "get_experiment_api_v1_experiments__experiment_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Experiment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "experiments"
        ],
        "summary": "Update Experiment",
        "description": "Update experiment (only status and metadata allowed).",
        "operationId": "update_experiment_api_v1_experiments__experiment_id__put",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Experiment Id"
            }
          },
          {
            "name": "status_update",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Update"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "type": "object",
                    "additionalProperties": true
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Metadata"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "experiments"
        ],
        "summary": "Delete Experiment",
        "description": "Delete an experiment.\n\nFirst cancels any running task, then deletes the experiment from database.",
        "operationId": "delete_experiment_api_v1_experiments__experiment_id__delete",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Experiment Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/experiments/{experiment_id}/repeat": {
      "post": {
        "tags": [
          "experiments"
        ],
        "summary": "Repeat Experiment",
        "description": "Repeat (duplicate) an existing experiment with the same configuration.\n\nCreates a new experiment with:\n- Same configuration (models, prompts, strategies, etc.)\n- New experiment_id\n- Name appended with \" (Repeat)\" or timestamp\n- Status set to \"pending\"",
        "operationId": "repeat_experiment_api_v1_experiments__experiment_id__repeat_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Experiment Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/experiments/{experiment_id}/iterations": {
      "get": {
        "tags": [
          "experiments"
        ],
        "summary": "Get Experiment Iterations",
        "description": "Get all iterations for an experiment with FULL details.\n\nReturns complete iteration data including:\n- Original and mutated prompts\n- Target response\n- Judge score and reasoning\n- Latency and metadata",
        "operationId": "get_experiment_iterations_api_v1_experiments__experiment_id__iterations_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Experiment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/experiments/{experiment_id}/statistics": {
      "get": {
        "tags": [
          "experiments"
        ],
        "summary": "Get Experiment Statistics",
        "description": "Get experiment statistics.",
        "operationId": "get_experiment_statistics_api_v1_experiments__experiment_id__statistics_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Experiment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExperimentConfig": {
        "properties": {
          "experiment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Experiment Id",
            "description": "Unique experiment ID"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "title": "Name",
            "description": "Experiment name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Experiment description"
          },
          "target_model_provider": {
            "$ref": "#/components/schemas/LLMProvider",
            "description": "Target model provider"
          },
          "target_model_name": {
            "type": "string",
            "title": "Target Model Name",
            "description": "Target model name"
          },
          "attacker_model_provider": {
            "$ref": "#/components/schemas/LLMProvider",
            "description": "Attacker model provider"
          },
          "attacker_model_name": {
            "type": "string",
            "title": "Attacker Model Name",
            "description": "Attacker model name"
          },
          "judge_model_provider": {
            "$ref": "#/components/schemas/LLMProvider",
            "description": "Judge model provider"
          },
          "judge_model_name": {
            "type": "string",
            "title": "Judge Model Name",
            "description": "Judge model name"
          },
          "initial_prompts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Initial Prompts",
            "description": "Initial prompts to test"
          },
          "strategies": {
            "items": {
              "$ref": "#/components/schemas/AttackStrategyType"
            },
            "type": "array",
            "minItems": 1,
            "title": "Strategies",
            "description": "Attack strategies to use"
          },
          "max_iterations": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Max Iterations",
            "description": "Maximum iterations",
            "default": 20
          },
          "max_concurrent_attacks": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Max Concurrent Attacks",
            "description": "Maximum concurrent attacks",
            "default": 5
          },
          "success_threshold": {
            "type": "number",
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Success Threshold",
            "description": "Success threshold score",
            "default": 7.0
          },
          "timeout_seconds": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Timeout Seconds",
            "description": "Experiment timeout in seconds",
            "default": 3600
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation timestamp"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata",
            "description": "Additional metadata"
          }
        },
        "type": "object",
        "required": [
          "name",
          "target_model_provider",
          "target_model_name",
          "attacker_model_provider",
          "attacker_model_name",
          "judge_model_provider",
          "judge_model_name",
          "initial_prompts",
          "strategies"
        ],
        "title": "ExperimentConfig",
        "description": "Experiment configuration model."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LLMProvider": {
        "type": "string",
        "enum": [
          "ollama",
          "azure",
          "openai"
        ],
        "title": "LLMProvider",
        "description": "LLM provider types."
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:9000/api/v1",
...
    },
    {
...
    }
  ]
}